+++ src/helpers/webview/bridge.js
@@ -4,9 +4,8 @@ import { popupContextMenu } from './context_menu';
 import { EVENT_WEBVIEW_NOTIFICATION, EVENT_NOTIFICATION_REFLECT_READY, EVENT_BRIDGE_INIT, EVENT_SPELLING_REFLECT_READY, EVENT_UPDATE_USER_SETTING } from '../../constants';
 import { ipcRenderer, remote } from 'electron';
 import InputManager from './input_manager';
-import { ENVIRONMENT } from 'hunspell-asm';
-import { SpellCheckerProvider } from 'electron-hunspell';
-
+import fs from 'fs';
+import { SpellCheckerProvider, attachSpellCheckProvider } from 'electron-hunspell';
 
 // Electron (or the build of Chromium it uses?) does not seem to have any default right-click menu, this adds our own.
 remote.getCurrentWebContents().addListener('context-menu', popupContextMenu);
@@ -17,7 +16,7 @@ window.onload = () => {
 
     // Without observing the DOM, we don't have a reliable way to let the main process know once
     // (and only once) that the main part of the app (not the QR code screen) has loaded, which is
-    // when we need to init the spellchecker and prompt Linux users for sudo
+    // when we need to init the spellchecker
     const onMutation = function (mutationsList, observer) {
         if (document.querySelector('mw-main-nav')) { // we're definitely logged-in if this is in the DOM
             ipcRenderer.send(EVENT_BRIDGE_INIT);
@@ -34,16 +33,25 @@ window.onload = () => {
 // and if so, (down)loads the necessary files, then sends an event to which the following listener responds and
 // loads the spellchecker, if needed.
 ipcRenderer.once(EVENT_SPELLING_REFLECT_READY, async (event, { dictionaryLocaleKey, spellCheckFiles, customWords }) => {
-    if (dictionaryLocaleKey && spellCheckFiles.userLanguageAffFile && spellCheckFiles.userLanguageDicFile) {
+    if (dictionaryLocaleKey && spellCheckFiles && spellCheckFiles.userLanguageAffFile && spellCheckFiles.userLanguageDicFile) {
         const provider = new SpellCheckerProvider();
         window.spellCheckHandler = provider;
-        await provider.initialize({ environment: ENVIRONMENT.NODE });
-        await window.spellCheckHandler.loadDictionary(dictionaryLocaleKey, spellCheckFiles.userLanguageDicFile,spellCheckFiles.userLanguageAffFile);
-        window.spellCheckHandler.switchDictionary(dictionaryLocaleKey);
-        if (window.spellCheckHandler.selectedDictionary in customWords) {
-            for (let i = 0, n = customWords[window.spellCheckHandler.selectedDictionary].length; i < n; i++) {
-                const word = customWords[window.spellCheckHandler.selectedDictionary][i];
-                window.spellCheckHandler.spellCheckerTable[window.spellCheckHandler.selectedDictionary].spellChecker.addWord(word);
+        await provider.initialize({}); // Empty brace correct, see: https://github.com/kwonoj/electron-hunspell/blob/master/example/browserWindow.ts
+
+        await provider.loadDictionary(
+            dictionaryLocaleKey,
+            fs.readFileSync(spellCheckFiles.userLanguageDicFile),
+            fs.readFileSync(spellCheckFiles.userLanguageAffFile)
+        );
+
+        const attached = await attachSpellCheckProvider(provider);
+        attached.switchLanguage(dictionaryLocaleKey);
+
+        let table = window.spellCheckHandler.spellCheckerTable;
+        if (dictionaryLocaleKey in customWords && table && dictionaryLocaleKey in table) {
+            for (let i = 0, n = customWords[dictionaryLocaleKey].length; i < n; i++) {
+                const word = customWords[dictionaryLocaleKey][i];
+                table[dictionaryLocaleKey].spellChecker.addWord(word);
             }
         }
     }